refactor : created utils, and move the code
